requirements.txt
Django==3.0.7
django-cors-headers==3.3.0
django-filter==2.3.0
django-mptt==0.11.0
djangorestframework==3.11.0
Pillow==7.1.2
psycopg2==2.8.5
dj-database-url
